Unify the way how we load xml content into the bpmn diagram
